+++ changelog.md
@@ -50,6 +50,7 @@
 - `getConfig()` and `getType()` are implemented as getter as well, resp.: `storage.config` and `storage.type`
 - The configuration object are no longer extensible; if you want to provide extra parameters you can use the `options` object, for instance:
 - removed adapter config types
+- added a member `configError` in storage that can be queried with a getter e.g. `storage.configError`
 
 ```typescript
  const conf: ConfigAmazonS3 = {
@@ -263,8 +264,8 @@ await b2.listBuckets();
 However, the bucket name that you've provided with the configuration url or object is available by calling `getConfig`:
 
 ```typescript
-const  s3 = new Storage("s3://key:secret@eu-west-2/bucketName");
-await s3.listFiles(s3.getConfig().bucketName, "your-file.jpg')
+const  s3 = new Storage("s3://key=key&secret=secret&region=eu-west-2&bucketName=erwe");
+await s3.listFiles(s3.config.bucketName, "your-file.jpg')
 ```
 
 # 1.4.7 - 1.5.2